requirements.txt
dill==0.3.5.1
from-root==1.0.2
gdown
flask-cors
flask

# YOLOv8 Requirements
hydra-core>=1.2.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
ultralytics==8.2.5

pandas>=1.1.4
seaborn>=0.11.0
setuptools>=65.5.1

-e .